cairo-draw: avoid leaking cairo state when reaching the element limit
Once we reach the limit of acquired elements - that is
rsvg_drawing_ctx_limits_exceeded returns TRUE - for some reason the
"affine" matrix on the stack becomes degenerate (probably zeros(N)). I
don't really understand why that is, given that the limits_exceeded
check *should* only be called in _rsvg_node_draw_children, after which
affine isn't touched anymore (nor is it available to draw_children).
Nontheless shouldn't a degenerate matrix prevent freeing up the cairo
resources. The erroneous behaviour was introdoced in f93a2ac which
claims "I've checked this quite carefully", but the previous behaviour
of not checking the output of not checking the result of
cairo_matrix_invert is functionally closer to this fix than it is to
erroneously skip freeing the cairo resources.

Related-to: a37f0f8