Don't parse strings to implement property default values
For convenience, a few of the `make_property!` macro invocations used
a `default` expression that used `parse_str()`.  This is so that the
initial value for the property, which is written in the specs as
strings, would be easy to cut&paste from the specs.

However, the values are simple enough that we can simply write
expressions for them, that don't need parsing.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1114>