Document: wrap the tree in a RefCell
Document::cascade needs a mutable tree and so requires &mut self, but
since that function was previously only called from the public API,
the public API could indeed have a mutable Document.  However, now
that we are going to do lazy cascading, the Document needs to be kept
immutable most of the time.  So, to get a mutable tree, we'll use RefCell.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1114>