fix handle_new_from_file api test
The test that verifies that files can be loaded using an url (file://)
constructs a url using relative path, using the current directory as
the base (e.g. file://./tests/bin/../../rsvg/tests/foo..svg).

Since the base is not "/", this makes this url invalid, and the test fails.

Instead of the relative path, determine the absolute path of the test file,
and construct the url from that.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>
Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1122>