Reduce the amount of code generated for the conversion of ParseError into ElementError
Found with cargo-llvm-lines.

Before:

  Lines                 Copies               Function name
  -----                 ------               -------------
    9976 (2.4%,  2.4%)     58 (0.5%,  0.5%)  <core::result::Result<O,cssparser::parser::ParseError<rsvg::error::ValueErrorKind>> as rsvg::error::AttributeResultExt<O>>::attribute::{{closure}}
     232 (0.1%, 53.8%)     58 (0.5%, 35.7%)  <core::result::Result<O,cssparser::parser::ParseError<rsvg::error::ValueErrorKind>> as rsvg::error::AttributeResultExt<O>>::attribute

  stripped binary size of rsvg-convert: 6579032

After:

  Lines                 Copies               Function name
  -----                 ------               -------------
     232 (0.1%, 52.6%)     58 (0.5%, 35.2%)  <core::result::Result<O,cssparser::parser::ParseError<rsvg::error::ValueErrorKind>> as rsvg::error::AttributeResultExt<O>>::attribute
     232 (0.1%, 52.7%)     58 (0.5%, 35.7%)  <core::result::Result<O,cssparser::parser::ParseError<rsvg::error::ValueErrorKind>> as rsvg::error::AttributeResultExt<O>>::attribute::{{closure}}

  stripped binary size of rsvg-convert: 6562648

No change in rsvg-bench timings for the adwaita icons.  After all,
this code path is for the error paths when parsing attributes, and
there are probably not many errors that exercise it in that data.

The change is just to grab a result.map_err(|e| { ... }) closure and
put it in a separate function.  The cause for the large amount of
generated code is that we have this:

    impl<'i, O> AttributeResultExt<O> for Result<O, ParseError<'i>> {
        fn attribute(self, attr: QualName) -> Result<O, ElementError> {
            self.map_err(|e| { ... })
        }
    }

The closure for .map_err() is monomorphized for all the Result<O,
ParseError<'i>>, but the E are always the same and of course O is not
used in the map_err() closure.  So, we just extract the closure into a
separate function and the resulting generated code for .map_err() is
much smaller.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1134>