(#1199): Fix generation of GObject Introspection data with respect to the pixbuf option
The answer to the question "is gdk-pixbuf enabled in the C API?" comes
from pixbuf_dep.found(), for example in the way
include/librsvg/meson.build uses it to define the LIBRSVG_HAVE_PIXBUF
macro for the generated header file.

When generating GObject Introspection data, we need to answer the same
question.  If gdk-pixbuf is enabled in the C API, that means we need
to tell meson's gnome.generate_gir() to include the GdkPixbuf-2.0
module as part of the API.

This all is different from the build_pixbuf_loader meson variable,
which just means that: should we build the (now obsolete) gdk-pixbuf
loader or not.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/1199

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1139>